JBPM-6918: Some values filled in the Multiple Subform don't get transfered to the next task. #2 - Fixed CDI Issues
