+++ kie-wb-common-forms/kie-wb-common-dynamic-forms/kie-wb-common-dynamic-forms-backend/src/main/java/org/kie/workbench/common/forms/dynamic/backend/server/context/generation/dynamic/impl/marshalling/FieldValueMarshallerRegistryImpl.java
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2018 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.kie.workbench.common.forms.dynamic.backend.server.context.generation.dynamic.impl.marshalling;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
+import java.util.function.Supplier;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.inject.Instance;
+import javax.inject.Inject;
+
+import org.kie.workbench.common.forms.model.FieldDefinition;
+
+@ApplicationScoped
+public class FieldValueMarshallerRegistryImpl implements FieldValueMarshallerRegistry {
+
+    private Map<Class<? extends FieldDefinition>, Supplier<FieldValueMarshaller>> marshallers = new HashMap<>();
+
+    @Inject
+    public FieldValueMarshallerRegistryImpl(Instance<FieldValueMarshaller<?, ?, ? extends FieldDefinition>> marshallerInstance) {
+        for (FieldValueMarshaller marshaller : marshallerInstance) {
+            marshallers.put(marshaller.getSupportedField(), marshaller.newInstanceSupplier());
+        }
+    }
+
+    @Override
+    public FieldValueMarshaller getMarshaller(FieldDefinition fieldDefinition) {
+        Optional<Supplier<FieldValueMarshaller>> optional = Optional.ofNullable(marshallers.get(fieldDefinition.getClass()));
+        if (optional.isPresent()) {
+            return optional.get().get();
+        }
+        return null;
+    }
+}
